+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased changes
+* Add OpenId connect handler for OAuth login controller
 
 ## v1.9.3 (14 apr 2021)
 * [1155](https://github.com/AmpersandTarski/Ampersand/issues/1155) Fix issue with MrgAtom exec-engine function in case of non-existing atoms